requirements/base.txt
django==4.1.1
python-dotenv==1.0.0
requests==2.31.0
PyJWT==2.8.0
django-cors-headers==4.3.1